Update HWTask1.py
The main idea of the program is: for a given word, the bk tree will find all the "similar" words, i.e. the ones with edit distance equal to 1. We then build an anytree that has as root the starting word and each node has as children the "similar" words,  so when we finally find v, the ending word, we can easilly find its absolute path.  In order to proceed through all the words in the dictionary from just one word, we created a list (words) to which we add , each time, all the new "similar" words.